feat: add i64 formatting to code

+++ src/endpoints/admin/custom/create_custom.rs
@@ -17,7 +17,7 @@ use jsonwebtoken::{Validation,Algorithm,decode,DecodingKey};
 
 
 pub_struct!(Deserialize; CreateCustom {
-    quest_id: u32,
+    quest_id: i64,
     name: String,
     desc: String,
     cta: String,
@@ -41,7 +41,7 @@ pub async fn handler(
     let quests_collection = state.db.collection::<QuestDocument>("quests");
 
 
-    let res= verify_quest_auth(user, &quests_collection, &(body.quest_id as i32)).await;
+    let res= verify_quest_auth(user, &quests_collection, &(body.quest_id as i64)).await;
     if !res {
         return get_error("Error creating task".to_string());
     };
